requirements.txt
yt-dlp>=2024.1.0
ffmpeg-python>=0.2.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai-whisper>=20231117

